card fixes and dealer face down card

+++ src/MonoBlackjack.Core/Players/Human.cs
@@ -2,9 +2,9 @@ namespace MonoBlackjack.Core.Players;
 
 public class Human : PlayerBase
 {
-    public int Bank { get; set; }
+    public decimal Bank { get; set; }
 
-    public Human(string name = "Player", int? startingBank = null)
+    public Human(string name = "Player", decimal? startingBank = null)
         : base(name)
     {
         Bank = startingBank ?? GameConfig.StartingBank;
